Using DS1307 module for RTC and keep clock

+++ src/AlarmClock.cpp
@@ -15,7 +15,7 @@ void AlarmClock::handle_alarm_clock() {
     alarmProcess.handle_alarm();
 }
 
-void AlarmClock::onSetDatetime(uint8_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) {
+void AlarmClock::onSetDatetime(uint16_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) {
     clockProcess.setDateTime(year, month, day, hours, minutes, seconds);
 }
