+++ src/AlarmObserver.h
@@ -7,7 +7,7 @@ class BluetoothManager;
 class AlarmObserver {
     public:
         void attachSubject(BluetoothManager *subject);
-        virtual void onSetDatetime(uint8_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) = 0;
+        virtual void onSetDatetime(uint16_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) = 0;
         virtual void onSetNapDuration(uint32_t durationInMillis) = 0;
 };
