+++ src/CommandObserver.h
@@ -8,7 +8,7 @@ class CommandParser;
 class CommandObserver {
     public:
         void attachSubject(CommandParser *parser);
-        virtual void onReceivedSetDatetimeCommand(uint8_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) = 0;
+        virtual void onReceivedSetDatetimeCommand(uint16_t year, uint8_t month, uint8_t day, uint8_t hours, uint8_t minutes, uint8_t seconds) = 0;
         virtual void onReceivedSetNapCommand(uint32_t durationInMillis) = 0;
 };
